Add normalizeIsUnique method to driver interface
We need this since there are differences on how a DB handles the unique column case
MySQL - MariaDB consider a column as unique if there is a unique index on the column, without the need for a unique constraint. So we need to adjust on how we resolve differences so that we take this into account (only for this DB type).