Feature type validation + SRID support
Specifying a feature type will pass it as a parameter to the underlying
spatial type, allowing PostGIS to validate that stored geometries match
expectations.

Specifying an SRID will do similarly, with the additional benefit of
allowing the coordinate system used for a geometry to be stored in order
to transform it into other coordinate systems.