adds an option to control whether columns are inserted by default
This adds an `insert` option, analogous to the existing `select`
and `readonly` options, for use in the `Column` decorator.  Just
like a column specified with `@Column({select: false})` will not
be selected by default in queries, and a column with
`@Column({readonly: true;})` will not be `UPDATE`d,
a column specified with `@Column({insert: false})`
will not be inserted by default in `INSERT`s.

I believe this fixes #85, and is relevant to (though not the same as)
the work in #2150.